projecthelmetdetection/requirements.txt
streamlit
opencv-python
numpy
Pillow
ultralytics
playsound